# backend/src/generated/java/api/openapi.yaml
openapi: 3.0.1
servers:
- description: Generated server url
  url: http://backend-patient-app-main.jonas-blum.ch
security:
- X-Coach-Key: []
paths:
...
  /patients/journal-entry-conversation/{conversationId}/messages:
    get:
      operationId: getAllMessages
      parameters:
      - explode: false
        in: path
        name: conversationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JournalConversationOutputDTO'
          description: OK
      tags:
      - journal-entry-conversation-controller
      x-accepts: '*/*'
    post:
      operationId: sendMessage
      parameters:
      - explode: false
        in: path
        name: conversationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalMessageDTO'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageOutputDTO'
          description: OK
      tags:
      - journal-entry-conversation-controller
      x-content-type: application/json
      x-accepts: '*/*'
...
  /patients/exercise-conversation/{conversationId}/messages:
    get:
      operationId: getAllMessages_2
      parameters:
      - explode: false
        in: path
        name: conversationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CompleteExerciseConversationOutputDTO'
          description: OK
      tags:
      - exercise-conversation-controller
      x-accepts: '*/*'
    post:
      operationId: sendMessage_1
      parameters:
      - explode: false
        in: path
        name: conversationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageDTO'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageOutputDTO'
          description: OK
      tags:
      - exercise-conversation-controller
      x-content-type: application/json
      x-accepts: '*/*'
  /patients/document-conversation/{conversationId}/messages:
    get:
      operationId: getAllMessages_4
      parameters:
      - explode: false
        in: path
        name: conversationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentConversationOutputDTO'
          description: OK
      tags:
      - document-conversation-controller
      x-accepts: '*/*'
    post:
      operationId: sendMessage_2
      parameters:
      - explode: false
        in: path
        name: conversationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageDTO'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageOutputDTO'
          description: OK
      tags:
      - document-conversation-controller
      x-content-type: application/json
      x-accepts: '*/*'
...
  /patients/conversations/messages/{conversationId}:
    get:
      operationId: getAllMessages_5
      parameters:
      - explode: false
        in: path
        name: conversationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CompleteConversationOutputDTO'
          description: OK
      tags:
      - conversation-controller
      x-accepts: '*/*'
    post:
      operationId: sendMessage_3
      parameters:
      - explode: false
        in: path
        name: conversationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageDTO'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageOutputDTO'
          description: OK
      tags:
      - conversation-controller
      x-content-type: application/json
      x-accepts: '*/*'
...
  /patients/journal-entries/{entryId}/chatbot:
    get:
      operationId: getAllMessages_1
      parameters:
      - explode: false
        in: path
        name: entryId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JournalChatbotOutputDTO'
          description: OK
      tags:
      - journal-entry-controller
      x-accepts: '*/*'
...
  /patients/documents/{documentId}/chatbot:
    get:
      operationId: getAllMessages_3
      parameters:
      - explode: false
        in: path
        name: documentId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentChatbotOutputDTO'
          description: OK
      tags:
      - document-controller
      x-accepts: '*/*'
...
  /patients/journal-entry-conversation/{conversationId}:
    delete:
      operationId: deleteJournalChat
      parameters:
      - explode: false
        in: path
        name: conversationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - journal-entry-conversation-controller
      x-accepts: application/json
...
components:
  schemas:
...
    CreateJournalMessageDTO:
      properties:
        message:
          type: string
        journalEntryTitle:
          type: string
        journalEntryContent:
          type: string
      type: object
    MessageOutputDTO:
      example:
        requestMessage: requestMessage
        conversationId: conversationId
        id: id
        responseMessage: responseMessage
        timestamp: timestamp
      properties:
        timestamp:
          type: string
        id:
          type: string
        conversationId:
          type: string
        responseMessage:
          type: string
        requestMessage:
          type: string
      type: object
...
    CreateMessageDTO:
      properties:
        message:
          type: string
      type: object
...
    JournalConversationOutputDTO:
      example:
        name: name
        messages:
        - requestMessage: requestMessage
          conversationId: conversationId
          id: id
          responseMessage: responseMessage
          timestamp: timestamp
        - requestMessage: requestMessage
          conversationId: conversationId
          id: id
          responseMessage: responseMessage
          timestamp: timestamp
        id: id
      properties:
        id:
          type: string
        name:
          type: string
        messages:
          items:
            $ref: '#/components/schemas/MessageOutputDTO'
          type: array
      type: object
...
    JournalChatbotOutputDTO:
      example:
        name: name
        id: id
      properties:
        name:
          type: string
        id:
          type: string
      type: object
...
    CompleteExerciseConversationOutputDTO:
      example:
        name: name
        messages:
        - requestMessage: requestMessage
          conversationId: conversationId
          id: id
          responseMessage: responseMessage
          timestamp: timestamp
        - requestMessage: requestMessage
          conversationId: conversationId
          id: id
          responseMessage: responseMessage
          timestamp: timestamp
        id: id
      properties:
        id:
          type: string
        name:
          type: string
        messages:
          items:
            $ref: '#/components/schemas/MessageOutputDTO'
          type: array
      type: object
    DocumentChatbotOutputDTO:
      example:
        welcomeMessage: welcomeMessage
        name: name
        id: id
      properties:
        name:
          type: string
        id:
          type: string
        welcomeMessage:
          type: string
      type: object
    DocumentConversationOutputDTO:
      example:
        name: name
        messages:
        - requestMessage: requestMessage
          conversationId: conversationId
          id: id
          responseMessage: responseMessage
          timestamp: timestamp
        - requestMessage: requestMessage
          conversationId: conversationId
          id: id
          responseMessage: responseMessage
          timestamp: timestamp
        id: id
      properties:
        id:
          type: string
        name:
          type: string
        messages:
          items:
            $ref: '#/components/schemas/MessageOutputDTO'
          type: array
      type: object
...
    CompleteConversationOutputDTO:
      example:
        shareWithAi: true
        welcomeMessage: welcomeMessage
        name: name
        messages:
        - requestMessage: requestMessage
          conversationId: conversationId
          id: id
          responseMessage: responseMessage
          timestamp: timestamp
        - requestMessage: requestMessage
          conversationId: conversationId
          id: id
          responseMessage: responseMessage
          timestamp: timestamp
        id: id
        shareWithCoach: true
        chatBotAvatar: ANIMALISTIC
      properties:
        id:
          type: string
        name:
          type: string
        welcomeMessage:
          type: string
        shareWithCoach:
          type: boolean
        shareWithAi:
          type: boolean
        messages:
          items:
            $ref: '#/components/schemas/MessageOutputDTO'
          type: array
        chatBotAvatar:
          enum:
          - ANIMALISTIC
          - BLOB
          - CRYSTAL
          - HUMANISTIC
          - NONE
          - PLANT
          - ROBOTIC
          - NEURALNETWORK
          type: string
      type: object
...
  securitySchemes:
    X-Coach-Key:
      description: Your coach’s API key
      in: header
      name: X-Coach-Key
      type: apiKey